+++ bmad/code/make_mat6.f90
@@ -86,7 +86,7 @@ if (.not. ele%is_on) mat6_calc_method = bmad_standard$
 if (mat6_calc_method == auto$) then
   select case (ele%tracking_method)
   case (bmad_standard$, linear$);   mat6_calc_method = bmad_standard$
-  case (custom$, fel_averaged$, fel_unaveraged$)
+  case (custom$)
                                     mat6_calc_method = custom$
   case (mad$);                      mat6_calc_method = mad$
   case (symp_lie_bmad$);            mat6_calc_method = symp_lie_bmad$