+++ bmad/code/track1.f90
@@ -261,7 +261,7 @@ case (symp_lie_ptc$)
 case (mad$)
   call track1_mad (end_orb, ele, param)
 
-case (custom$, fel_averaged$, fel_unaveraged$)
+case (custom$)
   if (.not. associated(track1_custom_ptr)) then
     call out_io (s_error$, r_name, 'TRACK1_CUSTOM_PTR HAS NOT BEEN SET IN THIS PROGRAM!', &
                                    'NEEDED FOR A CUSTOM ELEMENT, OR TRACKING_METHOD = CUSTOM,', &