+++ lucifer/code/fel_unaveraged_mod.f90
@@ -147,7 +147,7 @@ if (2 * ustate%l_ramp > l) then
 endif
 
 if (steps_per_period < 10) then
-  call out_io (s_error$, r_name, 'fel_steps_per_period BELOW THE CONVERGENCE FLOOR OF 10: \i0\ ', &
+  call out_io (s_error$, r_name, 'UNAVERAGED_STEPS_PER_PERIOD BELOW THE CONVERGENCE FLOOR OF 10: \i0\ ', &
                                  i_array = [steps_per_period])
   return
 endif